Basic Searching & Requirements
I felt like it was finally to commit after hours of not comitting. I have done a lot of things. First of all there is a basic search functionaility currently it freezes the app though and doesn't have icons, will fix later. In addition, I made a requirements text file and got rid of libraries we didn't need. ChatGPT also added adding files and directorys so I ran with it. I also got rid of some stuff in man that we didn't need anymore.